+++ src/AtomUI.Desktop.Controls/Spin/Spin.cs
@@ -2,7 +2,7 @@ using AtomUI.Controls.Commons;
 
 namespace AtomUI.Desktop.Controls;
 
-public class Spin : AbstractSpin
+public partial class Spin : AbstractSpin
 {
     public Spin()
     {