+++ src/AtomUI.Desktop.Controls/Spin/SpinIndicator.cs
@@ -2,7 +2,7 @@ using AtomUI.Controls.Commons;
 
 namespace AtomUI.Desktop.Controls;
 
-public class SpinIndicator : AbstractSpinIndicator
+public partial class SpinIndicator : AbstractSpinIndicator
 {
     public SpinIndicator()
     {